Update Programming Exercise 4-3.c
From Prata's description of the programming task it wasn't quite clear to me if the user should type in two different floating-point numbers and have them displayed in different output formats, or if this was to be a scanf conversion specifier exercise. Also, I didn't manage to have the exponential power value be displayed with three digits.